output works, input still broken
