+++ impsy/interaction_config.py
@@ -5,28 +5,13 @@ import time
 import datetime
 import numpy as np
 import queue
-import serial
 import tomllib
 from threading import Thread
-import mido
 import click
-from websockets.sync.server import serve
 from .utils import mdrnn_config
 import impsy.impsio as impsio
 
-
 np.set_printoptions(precision=2)
-
-
-def match_midi_port_to_list(port, port_list):
-    """Return the closest actual MIDI port name given a partial match and a list."""
-    if port in port_list:
-        return port
-    contains_list = [x for x in port_list if port in x]
-    if not contains_list:
-        return False
-    else:
-        return contains_list[0]
 
 
 def setup_logging(dimension, location = "logs/"):
@@ -38,17 +23,6 @@ def setup_logging(dimension, location = "logs/"):
                         level=logging.INFO,
                         format=log_format)
     click.secho(f'Logging enabled: {log_file}', fg='green')
-
-
-def open_raspberry_serial():
-    """Tries to open a serial port for MIDI IO on Raspberry Pi."""
-    try:
-        click.secho("Trying to open Raspberry Pi serial port for MIDI in/out.", fg='yellow')
-        ser = serial.Serial('/dev/ttyAMA0', baudrate=31250)
-    except:
-        ser = None
-        click.secho("Could not open serial port, might be in development mode.", fg='red')
-    return ser
 
 
 def build_network(sess, compute_graph, config):
@@ -77,192 +51,6 @@ def build_network(sess, compute_graph, config):
     return net
 
 
-class ImpsySender(object):
-    """Class for sending data to external clients via MIDI, WebSockets, Serial, or OSC."""
-
-    def __init__(self, config) -> None:
-        self.config = config # Config dictionary read in from config.toml
-        self.dimension = self.config["model"]["dimension"] # retrieve dimension from the config file.
-        self.last_midi_notes = {} # dict to store last played notes via midi
-        # MIDI port opening
-        click.secho("Opening MIDI port for input/output.", fg='yellow')
-        try:
-            desired_input_port = match_midi_port_to_list(self.config["midi"]["in_device"], mido.get_input_names())
-            self.midi_in_port = mido.open_input(desired_input_port)
-            click.secho(f"MIDI: in port is: {self.midi_in_port.name}", fg='green')
-        except: 
-            self.midi_in_port = None
-            click.secho("Could not open MIDI input.", fg='red')
-            click.secho(f"MIDI Inputs: {mido.get_input_names()}", fg = 'blue')
-        try:
-            desired_output_port = match_midi_port_to_list(self.config["midi"]["out_device"], mido.get_output_names())
-            self.midi_out_port = mido.open_output(desired_output_port)
-            click.secho(f"MIDI: out port is: {self.midi_out_port.name}", fg='green')
-        except:
-            self.midi_out_port = None
-            click.secho("Could not open MIDI output.", fg='red')
-            click.secho(f"MIDI Outputs: {mido.get_output_names()}", fg = 'blue')
-
-    def serial_send_midi(self, message):
-        """Sends a mido MIDI message via the very basic serial output on Raspberry Pi GPIO."""
-        try:
-            self.serial.write(message.bin())
-        except: 
-            pass
-
-    def send_midi_message(self, message):
-        """Send a MIDI message across all required outputs"""
-        # TODO: this is where we can have laggy performance, careful.
-        if self.midi_out_port is not None:
-            self.midi_out_port.send(message)
-        self.serial_send_midi(message)
-        self.websocket_send_midi(message)
-
-    def send_midi_note_on(self, channel, pitch, velocity):
-        """Send a MIDI note on (and implicitly handle note_off)"""
-        # stop the previous note
-        try:
-            midi_msg = mido.Message('note_off', channel=channel, note=self.last_midi_notes[channel], velocity=0)
-            self.send_midi_message(midi_msg)
-            # click.secho(f"MIDI: note_off: {self.last_midi_notes[channel]}: msg: {midi_msg.bin()}", fg="blue")
-            # do this by whatever other channels necessary
-        except KeyError:
-            click.secho("Something wrong with turning MIDI notes off!!", fg="red")
-            pass
-
-        # play the present note
-        midi_msg = mido.Message('note_on', channel=channel, note=pitch, velocity=velocity)
-        self.send_midi_message(midi_msg)
-        # click.secho(f"MIDI: note_on: {pitch}: msg: {midi_msg.bin()}", fg="blue")
-        self.last_midi_notes[channel] = pitch
-
-    def send_control_change(self, channel, control, value):
-        """Send a MIDI control change message"""
-        midi_msg = mido.Message('control_change', channel=channel, control=control, value=value)
-        self.send_midi_message(midi_msg)
-
-    def send_midi_note_offs(self):
-        """Sends note offs on any MIDI channels that have been used for notes."""
-        outconf = self.config["midi"]["output"]
-        out_channels = [x[1] for x in outconf if x[0] == "note_on"]
-        for i in out_channels:
-            try:
-                midi_msg = mido.Message('note_off', channel=i-1, note=self.last_midi_notes[i-1], velocity=0)
-                self.send_midi_message(midi_msg)
-                # click.secho(f"MIDI: note_off: {self.last_midi_notes[i-1]}: msg: {midi_msg.bin()}", fg="blue")
-            except KeyError:
-                click.secho("Something wrong with all MIDI Note off!", fg="red")
-                pass
-
-    def send_sound_command_midi(self, command_args):
-        """Sends sound commands via MIDI"""
-        assert len(command_args)+1 == self.dimension, "Dimension not same as prediction size." # Todo more useful error.
-        start_time = datetime.datetime.now()
-        outconf = self.config["midi"]["output"]
-        values = list(map(int, (np.ceil(command_args * 127))))
-        if VERBOSE:
-            click.secho(f'out: {values}', fg='green')
-
-        for i in range(self.dimension-1):
-            if outconf[i][0] == "note_on":
-                self.send_midi_note_on(outconf[i][1]-1, values[i], 127) # note decremented channel (0-15)
-            if outconf[i][0] == "control_change":
-                self.send_control_change(outconf[i][1]-1, outconf[i][2], values[i]) # note decrement channel (0-15)
-        duration_time = (datetime.datetime.now() - start_time).total_seconds()
-        if duration_time > 0.02:
-            click.secho(f"Sound command sending took a long time: {(duration_time):.3f}s", fg="red")
-        # TODO: is it a good idea to have all this indexing? easy to screw up.
-
-    def handle_midi_input(self):
-        """Handle MIDI input messages that might come from mido"""
-        if self.midi_in_port is None:
-            return # fail early if MIDI not open.
-        for message in self.midi_in_port.iter_pending():
-            if message.type == "note_on":
-                try:
-                    index = self.config["midi"]["input"].index(["note_on", message.channel+1])
-                    value = message.note / 127.0
-                    construct_input_list(index,value)
-                except ValueError:
-                    pass
-
-            if message.type == "control_change":
-                try:
-                    index = self.config["midi"]["input"].index(["control_change", message.channel+1, message.control])
-                    value = message.value / 127.0
-                    construct_input_list(index,value)
-                except ValueError:
-                    pass
-
-
-# TODO: some storage for all the output channels.
-WS_CLIENTS = set() # storage for potential ws clients.
-
-def websocket_send_midi(message):
-    """Sends a mido MIDI message via websockets if available."""
-    global ws_client
-
-    if message.type == "note_on":
-        ws_msg = f"/channel/{message.channel}/noteon/{message.note}/{message.velocity}"
-    elif message.type == "note_off":
-        ws_msg = f"/channel/{message.channel}/noteoff/{message.note}/{message.velocity}"
-    elif message.type == "control_change":
-        ws_msg = f"/channel/{message.channel}/cc/{message.control}/{message.value}"
-    else:
-        return
-    # click.secho(f"WS out: {ws_msg}")
-    # Broadcast the ws_msg to all clients (sync version can't use websockets.broadcast function so doing this naively)
-    for ws_client in WS_CLIENTS.copy():
-        try:
-            ws_client.send(ws_msg)
-        except:
-            WS_CLIENTS.remove(ws_client)
-
-
-def websocket_handler(websocket):
-    """Handle websocket input messages that might arrive"""
-    global WS_CLIENTS
-    WS_CLIENTS.add(websocket) # add websocket to the client list.
-    # do the actual handling
-    for message in websocket:
-        click.secho(f"WS: {message}", fg="red") # TODO: fine for debug, but should be removed really.
-        m = message.split('/')[1:]
-        msg_type = m[2]
-        chan = int(m[1]) # TODO: should this be chan+1 or -1 or something.
-        note = int(m[3])
-        vel = int(m[4])
-        if msg_type == "noteon":
-            # note_on
-            try:
-                index = config["midi"]["input"].index(["note_on", chan])
-                value = note / 127.0
-                construct_input_list(index,value)
-            except ValueError:
-                click.secho(f"WS in: exception with message {message}", fg="red")
-                pass
-        elif msg_type == "cc":
-            # cc
-            try:
-                index = config["midi"]["input"].index(["control_change", chan, note])
-                value = vel / 127.0
-                construct_input_list(index,value)
-            except ValueError:
-                click.secho(f"WS in: exception with message {message}", fg="red")
-                pass
-        # global websocket
-        # ws_msg = f"/channel/{message.channel}/noteon/{message.note}/{message.velocity}"
-        # ws_msg = f"/channel/{message.channel}/noteoff/{message.note}/{message.velocity}"
-        # ws_msg = f"/channel/{message.channel}/cc/{message.control}/{message.value}"
-
-
-def websocket_serve_loop():
-    """Threading websockets server following https://websockets.readthedocs.io/en/stable/reference/sync/server.html"""
-    hostname = config['websocket']['server_ip']
-    port = config['websocket']['server_port']
-    with serve(websocket_handler, hostname, port) as server:
-        server.serve_forever()
-
-
 class InteractionServer(object):
     """Interaction server class. Contains state and functions for the interaction loop."""
 
@@ -276,13 +64,15 @@ class InteractionServer(object):
         ## Load global variables from the config file.
         self.verbose = self.config["verbose"]
         self.dimension = self.config["model"]["dimension"] # retrieve dimension from the config file.
+        self.mode = self.config["interaction"]["mode"]
 
-        self.output_sender = ImpsySender(self.config)
+        ## Set up IO.
+        self.midi_sender = impsio.MIDIServer(self.config, self.construct_input_list)
+        self.midi_sender.connect()
+        self.websocket_sender = impsio.WebSocketServer(self.config, self.construct_input_list)
 
-        # Serial port opening
-        self.serial = open_raspberry_serial()
-
-        # Import Keras and tensorflow, doing this later to make CLI more responsive.
+        # Import Keras and tensorflow
+        ## TODO may be better to do this only when needed.
         click.secho("Importing MDRNN.", fg='yellow')
         start_import = time.time()
         import impsy.mdrnn as mdrnn
@@ -291,7 +81,6 @@ class InteractionServer(object):
 
         # Interaction Loop Parameters
         # All set to false before setting is chosen.
-        self.mode = self.config["interaction"]["mode"]
 
         # Interactive Mapping
         if self.mode == "callresponse":
@@ -324,14 +113,21 @@ class InteractionServer(object):
         self.interface_input_queue = queue.Queue()
         self.rnn_prediction_queue = queue.Queue()
         self.rnn_output_buffer = queue.Queue()
-        self.writing_queue = queue.Queue()
         self.last_user_interaction_time = time.time()
         self.last_user_interaction_data = mdrnn.random_sample(out_dim=self.dimension)
         self.rnn_prediction_queue.put_nowait(mdrnn.random_sample(out_dim=self.dimension))
         self.call_response_mode = 'call'
 
 
-    def construct_input_list(self, index, value):
+    def send_back_values(self, output_values):
+        """sends back sound commands to the MIDI/OSC/WebSockets outputs"""
+        output = np.minimum(np.maximum(output_values, 0), 1)
+        self.midi_sender.send(output)
+        self.websocket_sender.send(output)
+
+
+    # Todo this is the "callback" for our IO functions.
+    def construct_input_list(self, index: int, value: float) -> None:
         """constructs a dense input list from a sparse format (e.g., when receiving MIDI)
         """
         # set up dense interaction list
@@ -339,7 +135,7 @@ class InteractionServer(object):
         int_input[index] = value
         # log
         values = list(map(int, (np.ceil(int_input * 127))))
-        if VERBOSE:
+        if self.verbose:
             click.secho(f"in: {values}", fg='yellow')
         logging.info("{1},interface,{0}".format(','.join(map(str, int_input)),
                     datetime.datetime.now().isoformat()))
@@ -352,8 +148,9 @@ class InteractionServer(object):
         self.interface_input_queue.put_nowait(self.last_user_interaction_data)
         # Send values to output if in config
         if self.config["interaction"]["input_thru"]:
-                send_sound_command_midi(np.minimum(np.maximum(self.last_user_interaction_data[1:], 0), 1))
-
+                # This is where outputs are sent via impsio objects.
+                output_values = np.minimum(np.maximum(self.last_user_interaction_data[1:], 0), 1)
+                self.send_back_values(output_values)
 
     def make_prediction(self, sess, compute_graph, neural_net):
         """Part of the interaction loop: reads input, makes predictions, outputs results"""
@@ -409,10 +206,10 @@ class InteractionServer(object):
                     self.rnn_output_buffer.get()
                     self.rnn_output_buffer.task_done()
                 # close sound control over MIDI
-                send_midi_note_offs()
+                self.midi_sender.disconnect()
 
 
-    def playback_rnn_loop():
+    def playback_rnn_loop(self):
         """Plays back RNN notes from its buffer queue. This loop blocks and should run in a separate thread."""
         while True:
             item = self.rnn_output_buffer.get(block=True, timeout=None)  # Blocks until next item is available.
@@ -427,8 +224,8 @@ class InteractionServer(object):
             # put last played in queue for prediction.
             self.rnn_prediction_queue.put_nowait(np.concatenate([np.array([dt]), x_pred]))
             if self.rnn_to_sound:
-                # send_sound_command(x_pred)
-                send_sound_command_midi(x_pred)
+                # Send predictions to outputs via impsio objects
+                self.send_back_values(x_pred)
                 if self.config["log_predictions"]:
                     logging.info("{1},rnn,{0}".format(','.join(map(str, x_pred)),
                                 datetime.datetime.now().isoformat()))
@@ -462,8 +259,7 @@ class InteractionServer(object):
         # Threads
         click.secho("Preparing MDRNN thread.", fg='yellow')
         rnn_thread = Thread(target=self.playback_rnn_loop, name="rnn_player_thread", daemon=True)
-        click.secho("Preparing websocket thread.", fg='yellow')
-        ws_thread = Thread(target=websocket_serve_loop, name="ws_receiver_thread", daemon=True)
+        # self.websocket_sender.connect() # TODO verify websockets again.
 
         # Logging
         if self.config["log"]:
@@ -472,19 +268,19 @@ class InteractionServer(object):
         # Start threads and run IO loop
         try:
             rnn_thread.start()
-            ws_thread.start()
             click.secho("RNN Thread Started", fg="green")
             while True:
                 self.make_prediction(sess, compute_graph, net)
                 if self.config["interaction"]["mode"] == "callresponse":
-                    self.output_sender.handle_midi_input() # handles incoming midi queue
+                    self.midi_sender.handle_midi_input() # handles incoming midi queue
                     # TODO: handle other kinds of input here?
                     self.monitor_user_action()
         except KeyboardInterrupt:
             click.secho("\nCtrl-C received... exiting.", fg='red')
             rnn_thread.join(timeout=0.1)
-            ws_thread.join(timeout=0.1)
-            self.output_sender.send_midi_note_offs() # stop all midi notes.
+            
+            self.midi_sender.disconnect() # send note offs and disconnect
+            self.websocket_sender.disconnect()
         finally:
             click.secho("\nDone, shutting down.", fg='red')
